refactored project structure and adding custom function
