+++ src/interpreter/runtime/executor.cpp
@@ -1,6 +1,9 @@
+#include "mathlib.hpp"
 #include "runtime_base.hpp"
+#include <cassert>
+#include <vector>
 
-// TODO: add new instructions
+// TODO: consider raw_value in Idnt
 
 void executor(Frame& frame, const Utils::BC::Instruction& inst) {
   using namespace Utils;
@@ -8,13 +11,29 @@ void executor(Frame& frame, const Utils::BC::Instruction& inst) {
   using Operator = BC::Operator;
   using Idnt = BC::Idnt;
 
+  auto fstate = frame.get_state();
+  if(fstate.state == Frame::FState::decl) {
+    auto [fp_ok, m_fp] = frame.get_munit(frame.frame_id(), fstate.decl_id);
+    if(!fp_ok) {
+      frame.set_rt_result(RtResult::make_corrupted_error());
+      return;
+    }
+    auto fp = m_fp.get<Callable>();
+    if(!fp->is_valid()) {
+      frame.set_rt_result(RtResult::make_corrupted_error());
+      return;
+    }
+    fp->add_inst(inst);
+    return;
+  }
   if(inst.oper < Grammer::ALL_OPER_NAMES_LEN) {
     Debug::console << "Runtime: handling [" << Grammer::ALL_OPER_NAMES[inst.oper] << "] instruction\n";
   }
+  assert(fstate.state == Frame::FState::exec);
   switch(inst.oper) {
   case Operator::set: {
     Idnt dest = inst.idnts[0], source = inst.idnts[1];
-    auto [vs_ok, m_vs] = frame.get_munit(dest.frame_id, dest.idnt_id());
+    auto [vs_ok, m_vs] = frame.get_munit(source.frame_id, source.idnt_id());
     if(!vs_ok) {
       frame.set_rt_result(RtResult::make_corrupted_error());
       return;
@@ -46,6 +65,31 @@ void executor(Frame& frame, const Utils::BC::Instruction& inst) {
   HANDLE_BIN_INST(Operator::multiply, *)
   HANDLE_BIN_INST(Operator::minus, -)
   HANDLE_BIN_INST(Operator::divide, /)
+  case Operator::def: {
+    Idnt func_id = inst.idnts.front();
+    Callable cobj(inst.idnts.size() - 1);
+    auto [fp_ok, m_fp] = frame.get_munit(frame.frame_id(), func_id.idnt_id());
+    if(!fp_ok) {
+      frame.set_rt_result(RtResult::make_corrupted_error());
+      return;
+    }
+    m_fp.set(cobj);
+    frame.set_state(Frame::FState::decl, func_id.idnt_id());
+  }
+  case Operator::ret: {
+    Idnt idnt = inst.idnts[0];
+    auto [vi_ok, vi] = frame.get_munit(idnt.frame_id, idnt.idnt_id());
+    if(!vi_ok) {
+      frame.set_rt_result(RtResult::make_corrupted_error());
+      return;
+    }
+    auto [ppv_ok, ppv] = frame.get_munit(frame.pframe().frame_id(), -1);
+    if(!ppv_ok) {
+      frame.set_rt_result(RtResult::make_corrupted_error());
+      return;
+    }
+    ppv.set(vi.get<Object>());
+  }
   case Operator::func: {
     Idnt func = inst.idnts.back();
     auto [fp_ok, m_fp] = frame.get_munit(func.frame_id, func.idnt_id());
@@ -56,6 +100,10 @@ void executor(Frame& frame, const Utils::BC::Instruction& inst) {
     auto fp = m_fp.get<Callable>();
     if(!fp->is_valid()) {
       frame.emplace_rt_result(RtResult::InvalidUse, "Numbers are not callable object.");
+      return;
+    }
+    if(fp->arg_cnt() != inst.idnts.size() - 1) {
+      frame.emplace_rt_result(RtResult::InvalidUse, "Invalid use of function, expected ", fp->arg_cnt(), " arguments, found", inst.idnts.size() - 1);
       return;
     }
     auto pframe = frame.enter_new_frame();
@@ -84,6 +132,31 @@ void executor(Frame& frame, const Utils::BC::Instruction& inst) {
     ppv.set(ppv.get<Number>());
     break;
   }
+  case Operator::callbf: {
+    auto [pv_ok, m_pv] = frame.get_munit(frame.frame_id(), -1);
+    if(!pv_ok) {
+      frame.set_rt_result(RtResult::make_corrupted_error());
+      return;
+    }
+    auto fn = MathLangLib::builtin_fn.find(inst.idnts.back().idnt_str_const());
+    assert(fn != MathLangLib::builtin_fn.end());
+    Utils::DT::args_t arg_list;
+    for(auto it = std::next(inst.idnts.cbegin()); it != inst.idnts.cend(); ++it) {
+      auto [iv_ok, m_iv] = frame.get_munit(it->frame_id, it->idnt_id_const());
+      if(!iv_ok) {
+        frame.set_rt_result(RtResult::make_corrupted_error());
+        return;
+      }
+      auto iv = m_iv.get<Number>();
+      if(!iv->is_valid()) {
+        frame.emplace_rt_result(RtResult::InvalidUse, "Builtin functions can only call with Number-type.");
+        return;
+      }
+      arg_list.push_back(iv->cast_data<Utils::DT::number_t>());
+    }
+    auto fn_ret = (fn->second)(arg_list);
+    m_pv.set(Number(fn_ret));
+  }
   case Operator::print: {
     Idnt idnt = inst.idnts[0];
     auto [vi_ok, m_vi] = frame.get_munit(idnt.frame_id, idnt.idnt_id());
@@ -95,6 +168,8 @@ void executor(Frame& frame, const Utils::BC::Instruction& inst) {
     if(!vi->is_valid()) {
       if(frame.rt_result().code == RtResult::UndefinedVar)
         frame.emplace_rt_result(RtResult::UndefinedVar, "Cannot print an undefined value.");
+      else
+        frame.emplace_rt_result(RtResult::InvalidUse, "Functions cannot be printed.");
       return;
     }
     frame.emplace_rt_result(RtResult::Ok, Utils::String::to_string(*vi->cast_data<Number::number_t>()));
