+++ src/interpreter/runtime/runtime_base.hpp
@@ -1,13 +1,13 @@
 #ifndef RUNTIME_BASE_HPP
 #define RUNTIME_BASE_HPP
 
-#include "../utils.hpp"
+#include "utils/utils.hpp"
 #include <cstddef>
-#include <utility>
 #include <memory>
 #include <any>
 
-void executor(Frame&, const Utils::BC::InstList&);
+class Frame;
+void executor(Frame&, const Utils::BC::Instruction&);
 
 class Object;
 template<class T> using ObjPtr = std::shared_ptr<T>;
@@ -27,6 +27,7 @@ class MemUnit {
   MemUnit();
   template<class ObjT> const ObjPtr<ObjT> get() const;
   template<class ObjT> void set(const ObjT&) noexcept;
+  template<class ObjT> void set(const ObjPtr<ObjT>) noexcept;
 
   #ifdef DEBUG
     friend std::ostream& operator<<(std::ostream&, const RtMemUnit&);
@@ -58,12 +59,22 @@ class Frame {
   public:
   using RtResult = Utils::Pipline::RtResult;
   template<class T> using SafeRet = Utils::DT::SafeRet<T>;
+  struct FState {
+    enum State {
+      exec,
+      decl,
+    } state;
+    int decl_id;
+    FState() : state(exec), decl_id(-1) {}
+    FState(State _state, int _decl_id) : state(_state), decl_id(_decl_id) {}
+  };
 
   private:
   struct __Frame {
     int frame_id;
     std::shared_ptr<MemTable> mtable;
     std::shared_ptr<RtResult> rt_result;  // All related frame shares the same rt_result
+    FState fstate;
     _Frame pframe;
   };
 
@@ -83,9 +94,11 @@ class Frame {
   std::shared_ptr<MemTable> get_mtable();
   // alias for get_frame(frame_id).get_mtable()->get_munit(m_index)
   SafeRet<MemUnit> get_munit(const std::size_t frame_id, const int m_index);
-  const RtResult& rt_result();
+  RtResult const& rt_result();
   void set_rt_result(const RtResult&);
   template<class ...P> void emplace_rt_result(RtResult::ExitCode, P...);
+  FState get_state() const noexcept;
+  void set_state(FState::State, int def_id);
 };
 
 class Object : std::enable_shared_from_this<Object> {
@@ -99,16 +112,19 @@ class Object : std::enable_shared_from_this<Object> {
   public:
   Object();
   Object(data_p);
-  template<class T> std::shared_ptr<T> cast_data() const;
-  template<class ObjT> ObjPtr<ObjT> cast_self() const noexcept;
-  ObjPtr<Object> to_object() const;
+  template<class T> std::shared_ptr<T> cast_data();
+  template<class T> std::shared_ptr<T const> cast_data() const;
+  template<class ObjT> ObjPtr<ObjT> cast_self() noexcept;
+  template<class ObjT> ObjPtr<ObjT const> cast_self() const noexcept;
+  ObjPtr<Object> to_object();
+  ObjPtr<Object const> to_object() const;
 
   // For Object
   // Object type and its subtype cannot use method above
   // Only avaliable on ObjPtr<>
   //
   // return a new object holds by shared_ptr
-  ObjPtr<Object> to_ptr() const;
+  ObjPtr<Object> to_ptr() const noexcept;
   virtual bool is_valid() const { return true; };
 };
 
@@ -119,10 +135,14 @@ class Null : public Object {
 };
 
 class Callable : public Object {
+  protected:
+  int _arg_cnt;
+
   public:
   using InstList = Utils::BC::InstList;
+  using Instruction = Utils::BC::Instruction;
 
-  Callable();
+  Callable(int __arg_cnt);
   Callable(const InstList&);
   bool is_valid() const override;
   //  pframe->pre_value -> return value
@@ -130,7 +150,9 @@ class Callable : public Object {
   //  frame->mtable[1] -> arg[1]
   //  frame->mtable[2] -> arg[2]
   //  ...
-  void call(Frame&);
+  void call(Frame&) const;
+  void add_inst(const Instruction&);
+  int arg_cnt() const noexcept;
 };
 
 class Number : public Object {
@@ -145,5 +167,43 @@ class Number : public Object {
   Number operator*(const Number&);
   Number operator/(const Number&);
 };
+
+// template function implementation
+template<class ObjT> const ObjPtr<ObjT> MemUnit::get() const {
+  assert(obj_ptr != nullptr);
+  return obj_ptr->cast_self<ObjT>();
+}
+
+template<class ObjT> void MemUnit::set(const ObjT& value) noexcept {
+  obj_ptr = value.to_ptr();
+}
+
+template<class ObjT> void MemUnit::set(const ObjPtr<ObjT> value) noexcept {
+  obj_ptr = value;
+}
+
+template<class ...P> void Frame::emplace_rt_result(RtResult::ExitCode code, P... t) {
+  set_rt_result(RtResult(code, t...));
+}
+
+template<class T> std::shared_ptr<T> Object::cast_data() {
+  assert(data->type() == typeid(T));
+  return std::shared_ptr<T>(shared_from_this(), std::any_cast<T>(data.get())); 
+}
+
+template<class T> std::shared_ptr<T const> Object::cast_data() const {
+  assert(data->type() == typeid(T));
+  return std::shared_ptr<T>(shared_from_this(), std::any_cast<T>(data.get())); 
+}
+
+template<class ObjT> ObjPtr<ObjT> Object::cast_self() noexcept {
+  auto st = shared_from_this();
+  return ObjPtr<ObjT>(st, dynamic_cast<ObjT*>(st.get()));
+}
+
+template<class ObjT> ObjPtr<ObjT const> Object::cast_self() const noexcept {
+  auto st = shared_from_this();
+  return ObjPtr<ObjT const>(st, dynamic_cast<ObjT *const>(st.get()));
+}
 
 #endif
