+++ src/utils/utils.hpp
@@ -1,0 +1,14 @@
+#ifndef UTILS_HPP
+#define UTILS_HPP
+
+#include "./bc.hpp"
+#include "./cli.hpp"
+#include "./debug.hpp"
+#include "./dt.hpp"
+#include "./function.hpp"
+#include "./grammer.hpp"
+#include "./hash.hpp"
+#include "./pipline.hpp"
+#include "./string.hpp"
+
+#endif
