+++ tests/CMakeLists.txt
@@ -1,4 +1,4 @@
 add_executable(test_interpreter test_interpreter.cpp)
-target_link_libraries(test_interpreter MathLangInterpreter)
+target_link_libraries(test_interpreter Interpreter)
 
 add_test(NAME test_interpreter COMMAND test_interpreter)
